python-service/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
numpy
torch
torchvision
Pillow